ref: rename script files
