+++ index.html
@@ -230,8 +230,8 @@
 
   <div class="modal-window"></div>
 
-  <script src="./scripts/_data.js" type="module"></script>
-  <script src="./scripts/_slider.js" type="module"></script>
+  <script src="./scripts/data.js" type="module"></script>
+  <script src="./scripts/slider.js" type="module"></script>
   <script src="./scripts/main.js" type="module"></script>
 </body>
