+++ scripts/slider.js
